Project_Beachworm/requirements.txt
asgiref==3.2.3
certifi==2020.12.5
chardet==4.0.0
Django==3.0.3
django-cors-headers==3.7.0
django-environ==0.4.5
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
idna==2.10
PyJWT==2.0.1
pytz==2019.3
requests==2.25.1
six==1.15.0
spotipy==2.17.1
sqlparse==0.3.0
urllib3==1.26.3